[onert] Move remain ir test
This commit moves remain ir test into "src/" and remove test_onert unittest.
