+++ runtime/onert/core/src/ir/verifier/Verifier.test.cc
@@ -14,14 +14,15 @@
  * limitations under the License.
  */
 
+#include "Verifier.h"
+
+#include "../MockNode.h"
+
+#include "ir/Graph.h"
+
 #include <gtest/gtest.h>
 
-#include "ir/Operation.h"
-#include "ir/Graph.h"
-#include "ir/verifier/Verifier.h"
 #include <memory>
-#include "ir/Operand.h"
-#include "MockNode.h"
 
 using IndexSet = onert::ir::OperandIndexSequence;
 using Mock = onert_test::ir::SimpleMock;
